refactor(grown-ups-guide-menu-screen): rename props to types
